requirements.txt
python-dateutil==2.9.0.post0
Django==5.1.1
fontawesomefree==6.6.0
crispy-bootstrap5==2024.10
django-filter==25.1